requirements.txt
aiocontextvars==0.2.2
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorclass==2.2.0
colorlog==5.0.1
confluent-kafka==1.7.0
contextvars==2.4
croniter==1.0.15
dataclasses==0.8
elasticsearch==7.9.1
faust==1.10.4
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-bigquery==2.0.0
google-cloud-core==1.7.2
google-cloud-pubsub==2.1.0
google-cloud-secret-manager==1.0.0
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
haversine==2.3.0
idna==2.10
idna-ssl==1.1.0
immutables==0.16
kafka-python==1.4.7
libcst==0.3.19
mode==4.3.2
multidict==5.1.0
mypy-extensions==0.4.3
names==0.3.0
opentracing==1.3.0
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.0
robinhood-aiokafka==1.1.6
rsa==4.7.2
six==1.16.0
terminaltables==3.1.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
venusian==1.2.0
yarl==1.6.3
